sync

+++ application/Espo/Repositories/Attachment.php
@@ -31,6 +31,8 @@ namespace Espo\Repositories;
 
 use Espo\ORM\Entity;
 
+use Espo\Core\Utils\Util;
+
 class Attachment extends \Espo\Core\ORM\Repositories\RDB
 {
     protected function init()
@@ -69,14 +71,20 @@ class Attachment extends \Espo\Core\ORM\Repositories\RDB
     public function save(Entity $entity, array $options = array())
     {
         $isNew = $entity->isNew();
-        $result = parent::save($entity, $options);
 
         if ($isNew) {
+            $entity->id = Util::generateId();
+
             if (!empty($entity->id) && $entity->has('contents')) {
                 $contents = $entity->get('contents');
-                $this->getFileStorageManager()->putContents($entity, $contents);
+                $storeResult = $this->getFileStorageManager()->putContents($entity, $contents);
+                if ($storeResult === false) {
+                    throw new \Espo\Core\Exceptions\Error("Could not store the file");
+                }
             }
         }
+
+        $result = parent::save($entity, $options);
 
         return $result;
     }
@@ -84,7 +92,21 @@ class Attachment extends \Espo\Core\ORM\Repositories\RDB
     protected function afterRemove(Entity $entity, array $options = array())
     {
         parent::afterRemove($entity, $options);
-        $this->getFileStorageManager()->unlink($entity);
+
+        $duplicateCount = $this->where([
+            'OR' => [
+                [
+                    'sourceId' => $entity->getSourceId()
+                ],
+                [
+                    'id' => $entity->getSourceId()
+                ]
+            ],
+        ])->count();
+
+        if ($duplicateCount === 0) {
+            $this->getFileStorageManager()->unlink($entity);
+        }
     }
 
     public function getCopiedAttachment(Entity $entity, $role = null)
